Migrate regressions from 4001-4500
