+++ spacy/tests/lang/el/test_text.py
@@ -1,4 +1,15 @@
 import pytest
+from spacy.lang.el import Greek
+
+
+@pytest.mark.skip(reason="lemmatizer lookups no longer in vocab")
+@pytest.mark.issue(4272)
+def test_issue4272():
+    """Test that lookup table can be accessed from Token.lemma if no POS tags
+    are available."""
+    nlp = Greek()
+    doc = nlp("Χθες")
+    assert doc[0].lemma_
 
 
 def test_el_tokenizer_handles_long_text(el_tokenizer):
