+++ spacy/tests/serialize/test_serialize_pipeline.py
@@ -13,6 +13,8 @@ from spacy.pipeline.dep_parser import DEFAULT_PARSER_MODEL
 from spacy.pipeline.senter import DEFAULT_SENTER_MODEL
 from spacy.pipeline.tagger import DEFAULT_TAGGER_MODEL
 from spacy.pipeline.textcat import DEFAULT_SINGLE_TEXTCAT_MODEL
+from spacy.util import ensure_path, load_model
+from spacy.tokens import Span
 
 from ..util import make_tempdir
 
@@ -71,6 +73,7 @@ def test_issue3456():
     list(nlp.pipe(["hi", ""]))
 
 
+@pytest.mark.issue(3526)
 def test_issue_3526_1(en_vocab):
     patterns = [
         {"label": "HELLO", "pattern": "hello world"},
@@ -93,6 +96,7 @@ def test_issue_3526_1(en_vocab):
     assert new_ruler.ent_id_sep == ruler.ent_id_sep
 
 
+@pytest.mark.issue(3526)
 def test_issue_3526_2(en_vocab):
     patterns = [
         {"label": "HELLO", "pattern": "hello world"},
@@ -112,6 +116,7 @@ def test_issue_3526_2(en_vocab):
     assert new_ruler.overwrite is not ruler.overwrite
 
 
+@pytest.mark.issue(3526)
 def test_issue_3526_3(en_vocab):
     patterns = [
         {"label": "HELLO", "pattern": "hello world"},
@@ -132,6 +137,7 @@ def test_issue_3526_3(en_vocab):
         assert new_ruler.overwrite is not ruler.overwrite
 
 
+@pytest.mark.issue(3526)
 def test_issue_3526_4(en_vocab):
     nlp = Language(vocab=en_vocab)
     patterns = [{"label": "ORG", "pattern": "Apple"}]
@@ -147,6 +153,70 @@ def test_issue_3526_4(en_vocab):
         new_ruler = nlp2.get_pipe("entity_ruler")
         assert new_ruler.patterns == [{"label": "ORG", "pattern": "Apple"}]
         assert new_ruler.overwrite is True
+
+
+@pytest.mark.issue(4042)
+def test_issue4042():
+    """Test that serialization of an EntityRuler before NER works fine."""
+    nlp = English()
+    # add ner pipe
+    ner = nlp.add_pipe("ner")
+    ner.add_label("SOME_LABEL")
+    nlp.initialize()
+    # Add entity ruler
+    patterns = [
+        {"label": "MY_ORG", "pattern": "Apple"},
+        {"label": "MY_GPE", "pattern": [{"lower": "san"}, {"lower": "francisco"}]},
+    ]
+    # works fine with "after"
+    ruler = nlp.add_pipe("entity_ruler", before="ner")
+    ruler.add_patterns(patterns)
+    doc1 = nlp("What do you think about Apple ?")
+    assert doc1.ents[0].label_ == "MY_ORG"
+
+    with make_tempdir() as d:
+        output_dir = ensure_path(d)
+        if not output_dir.exists():
+            output_dir.mkdir()
+        nlp.to_disk(output_dir)
+        nlp2 = load_model(output_dir)
+        doc2 = nlp2("What do you think about Apple ?")
+        assert doc2.ents[0].label_ == "MY_ORG"
+
+
+@pytest.mark.issue(4042)
+def test_issue4042_bug2():
+    """
+    Test that serialization of an NER works fine when new labels were added.
+    This is the second bug of two bugs underlying the issue 4042.
+    """
+    nlp1 = English()
+    # add ner pipe
+    ner1 = nlp1.add_pipe("ner")
+    ner1.add_label("SOME_LABEL")
+    nlp1.initialize()
+    # add a new label to the doc
+    doc1 = nlp1("What do you think about Apple ?")
+    assert len(ner1.labels) == 1
+    assert "SOME_LABEL" in ner1.labels
+    apple_ent = Span(doc1, 5, 6, label="MY_ORG")
+    doc1.ents = list(doc1.ents) + [apple_ent]
+    # Add the label explicitly. Previously we didn't require this.
+    ner1.add_label("MY_ORG")
+    ner1(doc1)
+    assert len(ner1.labels) == 2
+    assert "SOME_LABEL" in ner1.labels
+    assert "MY_ORG" in ner1.labels
+    with make_tempdir() as d:
+        # assert IO goes fine
+        output_dir = ensure_path(d)
+        if not output_dir.exists():
+            output_dir.mkdir()
+        ner1.to_disk(output_dir)
+        config = {}
+        ner2 = nlp1.create_pipe("ner", config=config)
+        ner2.from_disk(output_dir)
+        assert len(ner2.labels) == 2
 
 
 @pytest.mark.parametrize("Parser", test_parsers)
